implement interactive xy correction

+++ multiTools/src/main/java/plugins/fmp/multitools/tools/imageTransform/CanvasImageTransformOptions.java
@@ -31,10 +31,10 @@ public class CanvasImageTransformOptions {
 	public int colordistanceType = 0;
 	public boolean ifGreater = true;
 
-	/** Pixel shift applied to the source before reference subtraction (drift preview). */
-	public int translateDx = 0;
-	/** Pixel shift applied to the source before reference subtraction (drift preview). */
-	public int translateDy = 0;
+	/** Pixel shift applied to the source before reference subtraction (drift preview; sub-pixel OK). */
+	public double translateDx = 0;
+	/** Pixel shift applied to the source before reference subtraction (drift preview; sub-pixel OK). */
+	public double translateDy = 0;
 
 	public final byte byteFALSE = 0;
 	public final byte byteTRUE = (byte) 0xFF;
